server/user/requirements.txt
grpcio-tools==1.39.0
googleapis-common-protos
pika